refactor(PreferencesForm): rozdziel logikę na mniejsze komponenty i custom hooki, popraw strukturę i czytelność kodu
